Update doc for year 2020
In 2020, HTTPS is ubiquotous, and fetching feeds over HTTP might not
even works because of redirects. So we now require clients to either
add our self-signed root certificate into their newsreader's store, or
just disable certificate checks.

(Yes, this is a doc update to vapourware that never got to the
development stage ¯\_(ツ)_/¯ )